improve audits overview
