+++ src/app/pages/auditor/experience/experience.component.html
@@ -43,9 +43,5 @@
             Close
             <mat-icon>close</mat-icon>
         </button>
-        <button mat-raised-button class="button-proximity" (click)="previewReport()">
-            View Full Report
-            <mat-icon>article</mat-icon>
-        </button>
     </section>
 </div>
